+++ sdk/formrecognizer/Azure.AI.FormRecognizer/src/RecognizeIdentityDocumentsOptions.cs
@@ -6,19 +6,19 @@ using System.Collections.Generic;
 namespace Azure.AI.FormRecognizer
 {
     /// <summary>
-    /// The set of options that can be specified when calling a Recognize ID Documents method
+    /// The set of options that can be specified when calling a Recognize identity Documents method
     /// to configure the behavior of the request. For example, specify the content type of the
     /// form, or whether or not to include form elements.
     /// </summary>
-    public class RecognizeIdDocumentsOptions
+    public class RecognizeIdentityDocumentsOptions
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="RecognizeIdDocumentsOptions"/> class which
-        /// allows to set options that can be specified when calling a Recognize ID Documents method
+        /// Initializes a new instance of the <see cref="RecognizeIdentityDocumentsOptions"/> class which
+        /// allows to set options that can be specified when calling a Recognize identity Documents method
         /// to configure the behavior of the request. For example, specify the content type of the
         /// form, or whether or not to include form elements.
         /// </summary>
-        public RecognizeIdDocumentsOptions()
+        public RecognizeIdentityDocumentsOptions()
         {
         }