+++ sdk/formrecognizer/Azure.AI.FormRecognizer/tests/samples/Sample14_RecognizeIdentityDocumentsFromFile.cs
@@ -15,14 +15,14 @@ namespace Azure.AI.FormRecognizer.Samples
     public partial class FormRecognizerSamples : SamplesBase<FormRecognizerTestEnvironment>
     {
         [Test]
-        public async Task RecognizeIdDocumentsFromFile()
+        public async Task RecognizeIdentityDocumentsFromFile()
         {
             string endpoint = TestEnvironment.Endpoint;
             string apiKey = TestEnvironment.ApiKey;
 
             FormRecognizerClient client = new FormRecognizerClient(new Uri(endpoint), new AzureKeyCredential(apiKey));
 
-            #region Snippet:FormRecognizerSampleRecognizeIdDocumentsFileStream
+            #region Snippet:FormRecognizerSampleRecognizeIdentityDocumentsFileStream
 #if SNIPPET
             string sourcePath = "<sourcePath>";
 #else
@@ -31,16 +31,16 @@ namespace Azure.AI.FormRecognizer.Samples
 
             using var stream = new FileStream(sourcePath, FileMode.Open);
 
-            RecognizeIdDocumentsOperation operation = await client.StartRecognizeIdDocumentsAsync(stream);
+            RecognizeIdentityDocumentsOperation operation = await client.StartRecognizeIdentityDocumentsAsync(stream);
             Response<RecognizedFormCollection> operationResponse = await operation.WaitForCompletionAsync();
-            RecognizedFormCollection idDocuments = operationResponse.Value;
+            RecognizedFormCollection identityDocuments = operationResponse.Value;
 
             // To see the list of all the supported fields returned by service and its corresponding types, consult:
             // https://aka.ms/formrecognizer/iddocumentfields
 
-            RecognizedForm idDocument = idDocuments.Single();
+            RecognizedForm identityDocument = identityDocuments.Single();
 
-            if (idDocument.Fields.TryGetValue("Address", out FormField addressField))
+            if (identityDocument.Fields.TryGetValue("Address", out FormField addressField))
             {
                 if (addressField.Value.ValueType == FieldValueType.String)
                 {
@@ -49,7 +49,7 @@ namespace Azure.AI.FormRecognizer.Samples
                 }
             }
 
-            if (idDocument.Fields.TryGetValue("Country", out FormField countryField))
+            if (identityDocument.Fields.TryGetValue("Country", out FormField countryField))
             {
                 if (countryField.Value.ValueType == FieldValueType.Country)
                 {
@@ -58,7 +58,7 @@ namespace Azure.AI.FormRecognizer.Samples
                 }
             }
 
-            if (idDocument.Fields.TryGetValue("DateOfBirth", out FormField dateOfBirthField))
+            if (identityDocument.Fields.TryGetValue("DateOfBirth", out FormField dateOfBirthField))
             {
                 if (dateOfBirthField.Value.ValueType == FieldValueType.Date)
                 {
@@ -67,7 +67,7 @@ namespace Azure.AI.FormRecognizer.Samples
                 }
             }
 
-            if (idDocument.Fields.TryGetValue("DateOfExpiration", out FormField dateOfExpirationField))
+            if (identityDocument.Fields.TryGetValue("DateOfExpiration", out FormField dateOfExpirationField))
             {
                 if (dateOfExpirationField.Value.ValueType == FieldValueType.Date)
                 {
@@ -76,7 +76,7 @@ namespace Azure.AI.FormRecognizer.Samples
                 }
             }
 
-            if (idDocument.Fields.TryGetValue("DocumentNumber", out FormField documentNumberField))
+            if (identityDocument.Fields.TryGetValue("DocumentNumber", out FormField documentNumberField))
             {
                 if (documentNumberField.Value.ValueType == FieldValueType.String)
                 {
@@ -85,7 +85,7 @@ namespace Azure.AI.FormRecognizer.Samples
                 }
             }
 
-            if (idDocument.Fields.TryGetValue("FirstName", out FormField firstNameField))
+            if (identityDocument.Fields.TryGetValue("FirstName", out FormField firstNameField))
             {
                 if (firstNameField.Value.ValueType == FieldValueType.String)
                 {
@@ -94,7 +94,7 @@ namespace Azure.AI.FormRecognizer.Samples
                 }
             }
 
-            if (idDocument.Fields.TryGetValue("LastName", out FormField lastNameField))
+            if (identityDocument.Fields.TryGetValue("LastName", out FormField lastNameField))
             {
                 if (lastNameField.Value.ValueType == FieldValueType.String)
                 {
@@ -103,7 +103,7 @@ namespace Azure.AI.FormRecognizer.Samples
                 }
             }
 
-            if (idDocument.Fields.TryGetValue("Region", out FormField regionfield))
+            if (identityDocument.Fields.TryGetValue("Region", out FormField regionfield))
             {
                 if (regionfield.Value.ValueType == FieldValueType.String)
                 {